Keep event history within maintainability limits
